Trait petri::crossover::CrossoverOp

pub trait CrossoverOp<G> {
    fn crossover<R: Rng>(&self, g1: &G, g2: &G, rng: &mut R) -> (G, G);
}

A crossover genetic operator

Required Methods

fn crossover<R: Rng>(&self, g1: &G, g2: &G, rng: &mut R) -> (G, G)

Return a two children of the given parent genomes

Panics

panics if the two given genomes differ in length