Struct petri::mutate::BitFlip

pub struct BitFlip { /* fields omitted */ }

A mutation genetic operator that will randomly invert the bits of each chromosome of a genome

Methods

impl BitFlip

pub fn with_pb(indv_pb: f32) -> Self

Create a Bitflip operator with a specified per chromosome mutation probability