Struct petri::runner::SimpleGARunner

pub struct SimpleGARunner<S, M, C, E, G, O, R> where
    S: SelectOp<G, O>,
    M: MutateOp<G>,
    C: CrossoverOp<G>,
    E: Fn(&G) -> O,
    G: Clone,
    O: Ord + Clone,
    R: Rng,  { /* fields omitted */ }

A Simple Genetic algorithm runner

Methods

impl<S, M, C, E, G, O, R> SimpleGARunner<S, M, C, E, G, O, R> where
S: SelectOp<G, O>,
M: MutateOp<G>,
C: CrossoverOp<G>,
E: Fn(&G) -> O,
G: Clone,
O: Ord + Clone + Zero,
R: Rng,

pub fn initialize_with_fn<F>(
select: S,
mutate: M,
crossover: C,
evaluate: E,
cx_pb: f32,
mut_pb: f32,
rng: R,
size: usize,
f: F
) -> Self where
F: Fn() -> G,

Create a new runner and intialize the initial population using a function

Panics

panics if either mut_pb or cx_pb are not a value between 0.0 and 1.0

pub fn initialize_with_state(
select: S,
mutate: M,
crossover: C,
evaluate: E,
cx_pb: f32,
mut_pb: f32,
rng: R,
state: State<G, O>
) -> Self

Create a new runner with a given internal state

Panics

panics if either mut_pb or cx_pb are not a value between 0.0 and 1.0

pub fn initialize_with_individual(
select: S,
mutate: M,
crossover: C,
evaluate: E,
cx_pb: f32,
mut_pb: f32,
rng: R,
size: usize,
individual: G
) -> Self

Create anew runner using an individiual to initialize the population

Panics

panics if either mut_pb or cx_pb are not a value between 0.0 and 1.0

pub fn fitnesses(&self) -> Vec<O>

Return the fitnesses of all individuals

pub fn population(&self) -> Vec<G>

Return the current population

pub fn state(&self) -> &State<G, O>

Returns a reference to the runnner state

pub fn size(&self) -> usize

Return the size of the population

pub fn generation(&self) -> usize

Return the current generation number

pub fn advance(&mut self)

Advance the algorithm state one generation